Added original Pan For Gold Project
Hopefully this is the proper file. This was my old version of Pan For Gold without any method implementation, this shit is so ass ngl. Previous versions have been lost to time, however this version is the best existing version out of them so its not a big deal I guess.